+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/domain/models/Album.kt
@@ -3,9 +3,9 @@ package com.example.kmmtutorial.domain.models
 
 @kotlinx.serialization.Serializable
 data class Album(
-    val id: Int,
+    val id: Long,
     var previewPhoto: String,
     val title: String,
-    val userId: Int,
+    val userId: Long,
     var username: String,
 )
