+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/domain/models/Comment.kt
@@ -4,7 +4,7 @@ package com.example.kmmtutorial.domain.models
 data class Comment(
     val body: String,
     val email: String,
-    val id: Int,
+    val id: Long,
     val name: String,
-    val postId: Int
+    val postId: Long
 )
