+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/domain/models/Photo.kt
@@ -3,8 +3,8 @@ package com.example.kmmtutorial.domain.models
 
 @kotlinx.serialization.Serializable
 data class Photo(
-    val albumId: Int,
-    val id: Int,
+    val albumId: Long,
+    val id: Long,
     val thumbnailUrl: String,
     val title: String,
     val url: String
