+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/domain/repositories/CommentsRepository.kt
@@ -4,5 +4,5 @@ import com.example.kmmtutorial.domain.models.Comment
 
 interface CommentsRepository {
     @Throws(Exception::class)
-    suspend fun getPostComments(postId: Int): List<Comment>
+    suspend fun getPostComments(postId: Long): List<Comment>
 }
