+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/domain/repositories/UserRepository.kt
@@ -4,5 +4,5 @@ import com.example.kmmtutorial.domain.models.User
 
 interface UserRepository {
     @Throws(Exception::class)
-    suspend fun getUser(userId: Int): User
+    suspend fun getUser(userId: Long): User
 }
